+++ src/index.js
@@ -20,7 +20,6 @@ import './js/frontend/head2head.js';
 import './js/frontend/teams.js';
 import './js/frontend/regulations.js';
 import './js/frontend/dragFile.js';
-import './js/frontend/news.js'
 import './js/frontend/devTools.js';
 
 import './themes.css'